+++ packages/cli-api/src/index.ts
@@ -1,6 +1,9 @@
 export { type Args, type Flags, BaseCommand } from "./baseCommand.js";
 export { CommandWithConfig, CommandWithoutConfig } from "./configCommand.js";
-export { RegExpFlag } from "./flags.js";
+export {
+	RegExpFlag,
+	ConfigFileFlag,
+} from "./flags.js";
 export type { CommitMergeability } from "./git.js";
 export { findGitRoot } from "./git.js";
 export { GitCommand } from "./gitCommand.js";
@@ -11,15 +14,14 @@ export {
 	revList,
 	shortCommit,
 } from "./git.js";
+export {
+	readJsonWithIndent,
+	updatePackageJsonFile,
+	type JsonWriteOptions,
+	type PackageTransformer,
+} from "./json.js";
 export type {
 	ErrorLoggingFunction,
 	LoggingFunction,
 	Logger,
 } from "./logger.js";
-export {
-	isSorted,
-	sortTsconfigFile,
-	type OrderList,
-	type SortTsconfigResult,
-	TsConfigSorter,
-} from "./tsconfig.js";