+++ packages/fundamentals/test/set.test.ts
@@ -2,34 +2,36 @@ import { describe, expect, it } from "vitest";
 
 import { addAll } from "../src/set.js";
 
-describe("addAll", () => {
-	it("add from array", () => {
-		const input = [1, 2, 3];
-		const theSet: Set<number> = new Set();
-		const actual = [...addAll(theSet, input)];
-		for (const item of input) {
-			expect(actual).toContain(item);
-		}
-		expect(actual).toHaveLength(3);
-	});
+describe("set APIs", () => {
+	describe("addAll", () => {
+		it("add from array", () => {
+			const input = [1, 2, 3];
+			const theSet: Set<number> = new Set();
+			const actual = [...addAll(theSet, input)];
+			for (const item of input) {
+				expect(actual).toContain(item);
+			}
+			expect(actual).toHaveLength(3);
+		});
 
-	it("add from generator", () => {
-		const input = range(1, 5);
-		const theSet: Set<number> = new Set();
-		const actual = [...addAll(theSet, input)];
-		for (const item of input) {
-			expect(actual).toContain(item);
-		}
-		expect(actual).toHaveLength(5);
-	});
+		it("add from generator", () => {
+			const input = range(1, 5);
+			const theSet: Set<number> = new Set();
+			const actual = [...addAll(theSet, input)];
+			for (const item of input) {
+				expect(actual).toContain(item);
+			}
+			expect(actual).toHaveLength(5);
+		});
 
-	it("handles duplicate inputs", () => {
-		const theSet: Set<number> = new Set(range(1, 5));
-		const actual = [...addAll(theSet, range(0, 3))];
-		for (const item of range(0, 5)) {
-			expect(actual).toContain(item);
-		}
-		expect(actual).toHaveLength(6);
+		it("handles duplicate inputs", () => {
+			const theSet: Set<number> = new Set(range(1, 5));
+			const actual = [...addAll(theSet, range(0, 3))];
+			for (const item of range(0, 5)) {
+				expect(actual).toContain(item);
+			}
+			expect(actual).toHaveLength(6);
+		});
 	});
 });