+++ syncpack.config.cjs
@@ -13,7 +13,14 @@ const config = {
 			strategy: "name@version",
 		},
 	},
-	dependencyTypes: ["dev", "engines", "packageManager", "peer", "prod"],
+	dependencyTypes: [
+		"dev",
+		"engines",
+		// Disabled for now because it removes the sha from the field.
+		// "packageManager",
+		"peer",
+		"prod",
+	],
 	versionGroups: [
 		{
 			label: "Use workspace protocol for workspace dependencies",
@@ -23,12 +30,13 @@ const config = {
 		},
 	],
 	semverGroups: [
-		{
-			label: "Use exact ranges for packageManager field",
-			range: "",
-			dependencyTypes: ["packageManager"],
-			packages: ["**"],
-		},
+		// Disabled for now because it removes the sha from the field.
+		// {
+		// 	label: "Use exact ranges for packageManager field",
+		// 	range: "",
+		// 	dependencyTypes: ["packageManager"],
+		// 	packages: ["**"],
+		// },
 		{
 			label: "Use exact ranges for these deps",
 			range: "",