+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @tylerbu/cli
 
+## 0.3.2
+
+### Patch Changes
+
+<details><summary>Updated 2 dependencies</summary>
+
+<small>
+
+[`cbdec3f`](https://github.com/tylerbutler/tools-monorepo/commit/cbdec3f7b3daa4ec642b44a5de046fff8420f15a) [`d55c982`](https://github.com/tylerbutler/tools-monorepo/commit/d55c982f960b56a79f0e0d35dd9102a25882032f)
+
+</small>
+
+- `@tylerbu/cli-api@0.4.0`
+- `dill-cli@0.1.2`
+
+</details>
+
+## 0.3.1
+
+### Patch Changes
+
+- Fix homepage URL _[`#99`](https://github.com/tylerbutler/tools-monorepo/pull/99) [`0654323`](https://github.com/tylerbutler/tools-monorepo/commit/06543231947fa5267863e5467d5837a51cf3d44b) [@tylerbutler](https://github.com/tylerbutler)_
+
+<details><summary>Updated 2 dependencies</summary>
+
+<small>
+
+[`0654323`](https://github.com/tylerbutler/tools-monorepo/commit/06543231947fa5267863e5467d5837a51cf3d44b)
+
+</small>
+
+- `@tylerbu/cli-api@0.3.1`
+- `dill-cli@0.1.1`
+
+</details>
+
 ## 0.3.0
 
 ### Minor Changes