+++ packages/fundamentals/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tylerbu/fundamentals
 
+## 0.1.1
+
+### Patch Changes
+
+- Fix homepage URL _[`#99`](https://github.com/tylerbutler/tools-monorepo/pull/99) [`0654323`](https://github.com/tylerbutler/tools-monorepo/commit/06543231947fa5267863e5467d5837a51cf3d44b) [@tylerbutler](https://github.com/tylerbutler)_
+
 ## 0.1.0
 
 ### Minor Changes